WIP on link to holdings from hitlist
