+++ lxl-web/src/routes/(app)/[[lang=lang]]/[fnurgel=fnurgel]/+page.server.ts
@@ -90,6 +90,8 @@ export const load = async ({ params, url, locals, fetch }) => {
 
 	const itemLinksByBibId = getItemLinksByBibId(bibIdsByInstanceId, locale, displayUtil);
 
+	console.log('itemLinksByBibId', itemLinksByBibId);
+
 	return {
 		type: mainEntity[JsonLd.TYPE],
 		types: types,
